+++ src/controllers/AssetsController.php
@@ -1372,6 +1372,23 @@ class AssetsController extends Controller
         $folder = $asset->getFolder();
         $sourcePath[] = $folder->getSourcePathInfo();
 
+        // for a JSON response (e.g. via element actions)
+        if ($this->request->getAcceptsJson()) {
+            // get all the way up to the root folder, cause we need source path info for each step
+            while (($parent = $folder->getParent()) !== null) {
+                $sourcePath[] = $parent->getSourcePathInfo();
+                $folder = $parent;
+            }
+
+            $data = [
+                'filename' => $asset->filename,
+                'sourcePath' => array_reverse($sourcePath),
+            ];
+
+            return $this->asJson($data);
+        }
+
+        // for a redirect response (e.g. element action menu items)
         $uri = StringHelper::ensureLeft(UrlHelper::prependCpTrigger($sourcePath[0]['uri']), '/');
         $url = UrlHelper::urlWithParams($uri, [
             'search' => $asset->filename,